+++ src/setup.jl
@@ -56,6 +56,7 @@ export grid_section
 export handle_qc
 export interpolate_barnes
 export label_from_varname
+export MLD_CF
 export N2
 export N2_spline
 export N2_first_difference